Upgrade requirements.txt dependencies
We were dependent on the requests library version 2.12.0, which is
vulnerable to CVE-2018-18074. This upgrades us to a version with a fix.

I regenerated the `requirements.txt` file by running the `pip-compile`
command it mentioned. This did not result in any changes, so I removed
the file, then reran the command. This time multiple changes were made,
including an upgrade of the `requests` package to a safe version.